requirements.txt
amqp==5.0.2
appdirs==1.4.3
arabic-reshaper==2.1.1
asgiref==3.3.1
attrs==20.3.0
Automat==0.8.0
backcall==0.1.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloud-init==20.4
colorama==0.4.3
command-not-found==0.3
conda==4.3.16
configobj==5.0.6
constantly==15.1.0
cryptography==3.3.1
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.5
django-allauth==0.44.0
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2
django-extensions==3.1.1
django-filter==2.4.0
django-heroku==0.3.1
django-markdown-deux==1.0.5
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-pagedown==2.2.0
django-semanticui-forms==1.6.5
django-unused-media==0.2.2
djangorestframework==3.12.2
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
html-text==0.5.2
html5lib==1.1
httpie==1.0.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.5.0
incremental==16.10.1
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.10.1
Js2Py==0.71
jsonfield==3.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.6.2
Markdown==3.3.4
markdown2==2.3.10
MarkupSafe==1.1.0
matplotlib==3.4.1
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.19.5
oauthlib==3.1.0
openpyxl==3.0.7
packaging==21.0
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.1.0
pipenv==11.9.0
pipwin==0.5.1
prompt-toolkit==3.0.10
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycosat==0.6.3
pycparser==2.20
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyjsparser==2.7.1
PyJWT==2.0.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.1
PyPrind==2.11.3
pyrsistent==0.17.3
pyserial==3.4
pySmartDL==1.3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.4
python3-openid==3.2.0
pyttsx3==2.90
pytz==2020.5
PyYAML==5.3.1
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
sos==4.0
soupsieve==2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
ssh-import-id==5.10
swapper==1.1.2.post1
systemd-python==234
traitlets==4.3.3
Twisted==18.9.0
tzlocal==2.1
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
vine==5.0.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
zipp==1.0.0
zope.interface==4.7.1
